🐛 Fix BlocklyApp props type definition
- ⚙️ Refactor some repetitive test code.
- 🚀 Add the randomizers helper.
- 📝 Improve documentation.
