+++ core/engines/pixijs/engineGenerator/index.stories.jsx
@@ -22,19 +22,11 @@ import {
   MOVE_LEFT_MOCK,
 } from 'core/__mocks__/InstructionsMock';
 import gameMetadata from 'core/__mocks__/gameMetadata';
-import componentGenerator from 'core/engines/pixijs/components/componentGenerator';
-import positioningFunctionalityBuilder
-  from 'core/engines/pixijs/components/functionalities/positioningFunctionalityBuilder';
+import basicActorGenerator from 'core/__mocks__/basicActorGenerator';
 
 import positioningProcessorBuilder from './processors/positioningProcessorBuilder';
 
 import engineGenerator from '.';
-
-const basicActorGenerator = (size: number, margin: number) =>
-  (view: Container, position: string) => (
-    componentGenerator(view, size, margin)
-        .addFunctionality('positioning', positioningFunctionalityBuilder(position))
-  );
 
 storiesOf('engines/pixijs/engineGenerator', module)
     .add('Simple engine with positioning', () => {
