+++ core/helpers/randomizers.js
@@ -1,0 +1,14 @@
+/**
+ * @author Luciano Graziani @lgraziani2712
+ * @license {@link http://www.opensource.org/licenses/mit-license.php|MIT License}
+ *
+ * @flow
+ */
+
+export function getRandomInt(min: number, max: number) {
+  return Math.floor(Math.random() * (max - min)) + min;
+}
+
+export function getRandomFloat(min: number, max: number) {
+  return Math.random() * (max - min) + min;
+}
